% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{UDG.datasets}
\alias{UDG.datasets}
\title{Show UDG datasets}
\usage{
UDG.datasets(pattern = "", full.info = FALSE)
}
\arguments{
\item{pattern}{Optional. Pattern in the dataset name as passed to function \code{\link{grep}} (case-insensitive).}

\item{full.info}{Default to FALSE. If TRUE, the complete URL is also displayed.}
}
\value{
The datasets table, in the form of a \code{data.frame}
}
\description{
Lists the harmonized datasets that are accessible via UDG.
}
\note{
The function assumes that the user has read permission to the package installation directory
}
\examples{
# Default built-in datasets
str(UDG.datasets())
(sets <- UDG.datasets())
names(sets)
sets$OBSERVATIONS
# Using argument pattern
UDG.datasets(pattern = "CORDEX-EUR44.*historical")
# Showing the full Url of the datasets
UDG.datasets(pattern = "E-OBS", full.info = TRUE)
}
\author{
M Iturbide
}
